Allow using GradedAxes and Sectors with BlockSparseArrays
