+++ NDTensors/src/NDTensors.jl
@@ -28,6 +28,7 @@ for lib in [
   :BroadcastMapConversion,
   :Unwrap,
   :RankFactorization,
+  :Sectors,
   :GradedAxes,
   :TensorAlgebra,
   :SparseArrayInterface,
@@ -36,7 +37,6 @@ for lib in [
   :BlockSparseArrays,
   :NamedDimsArrays,
   :SmallVectors,
-  :Sectors,
   :SortedSets,
   :TagSets,
   :UnallocatedArrays,
