+++ NDTensors/src/lib/GradedAxes/ext/GradedAxesSectorsExt/Project.toml
@@ -1,0 +1,1 @@
+[deps]
