+++ NDTensors/src/lib/GradedAxes/src/GradedAxes.jl
@@ -3,4 +3,5 @@ include("groupsortperm.jl")
 include("tensor_product.jl")
 include("abstractgradedunitrange.jl")
 include("gradedunitrange.jl")
+include("../ext/GradedAxesSectorsExt/src/GradedAxesSectorsExt.jl")
 end
