+++ NDTensors/src/lib/Sectors/src/Sectors.jl
@@ -1,8 +1,13 @@
 module Sectors
 
 include("abstractcategory.jl")
-include("group_definitions.jl")
-include("category_definitions.jl")
+include("category_definitions/u1.jl")
+include("category_definitions/zn.jl")
+include("category_definitions/su.jl")
+include("category_definitions/su2.jl")
+include("category_definitions/su2k.jl")
+include("category_definitions/fib.jl")
+include("category_definitions/ising.jl")
 
 include("namedtuple_operations.jl")
 include("category_product.jl")
