+++ NDTensors/src/lib/Sectors/src/abstractcategory.jl
@@ -6,35 +6,38 @@ function dimension(c::AbstractCategory)
   return error("method `dimension` not defined for type $(typeof(c))")
 end
 
-function label_fusion_rule(::Type{C}, l1, l2) where {C<:AbstractCategory}
-  return error("label_fusion_rule not defined for type $(C)")
+function label_fusion_rule(category_type::Type{<:AbstractCategory}, l1, l2)
+  return error("`label_fusion_rule` not defined for type $(category_type).")
 end
 
 function fusion_rule(c1::AbstractCategory, c2::AbstractCategory)
-  C = typeof(c1)
-  return [C(d) for d in label_fusion_rule(C, label(c1), label(c2))]
+  category_type = typeof(c1)
+  return [category_type(l) for l in label_fusion_rule(category_type, label(c1), label(c2))]
 end
 
 ⊗(c1::AbstractCategory, c2::AbstractCategory) = fusion_rule(c1, c2)
 
-Base.:(*)(s1::AbstractCategory, s2::AbstractCategory) = ⊗(s1, s2)
+⊕(c1::AbstractCategory, c2::AbstractCategory) = [c1, c2]
+⊕(cs::Vector{<:AbstractCategory}, c::AbstractCategory) = [cs; c]
+⊕(c::AbstractCategory, cs::Vector{<:AbstractCategory}) = [c; cs]
 
-⊕(a::AbstractCategory, b::AbstractCategory) = [a, b]
-⊕(v::Vector{<:AbstractCategory}, b::AbstractCategory) = vcat(v, b)
-⊕(a::AbstractCategory, v::Vector{<:AbstractCategory}) = vcat(a, v)
-
-function Base.show(io::IO, q::Vector{<:AbstractCategory})
-  (length(q) <= 1) && print(io, "[")
+function Base.show(io::IO, cs::Vector{<:AbstractCategory})
+  (length(cs) <= 1) && print(io, "[")
   symbol = ""
-  for l in q
-    print(io, symbol, l)
+  for c in cs
+    print(io, symbol, c)
     symbol = " ⊕ "
   end
-  return (length(q) <= 1) && print(io, "]")
+  (length(cs) <= 1) && print(io, "]")
+  return nothing
 end
 
-function trivial(::Type{C}) where {C<:AbstractCategory}
-  return error("method `trivial` not defined for type $C")
+function trivial(category_type::Type{<:AbstractCategory})
+  return error("`trivial` not defined for type $(category_type).")
 end
 
-istrivial(C::AbstractCategory) = (C == trivial(typeof(C)))
+istrivial(c::AbstractCategory) = (c == trivial(typeof(c)))
+
+dual(c::AbstractCategory) = typeof(c)(-label(c))
+
+Base.isless(c1::AbstractCategory, c2::AbstractCategory) = isless(label(c1), label(c2))
