+++ NDTensors/src/lib/Sectors/src/category_definitions/su2k.jl
@@ -1,0 +1,19 @@
+using HalfIntegers: Half
+
+#
+# Quantum 'group' su2ₖ
+#
+
+struct su2{k} <: AbstractCategory
+  j::Half{Int}
+end
+
+label(s::su2) = s.j
+
+level(s::su2{k}) where {k} = k
+
+trivial(::Type{su2{k}}) where {k} = su2{k}(0)
+
+function label_fusion_rule(::Type{su2{k}}, j1, j2) where {k}
+  return abs(j1 - j2):min(k - j1 - j2, j1 + j2)
+end
